+++ publish/CHANGELOG.md
@@ -1,5 +1,6 @@
 - v1.35
   - Improves suppport for massive locations (radius larger than 32 meters).
+  - Fixes the water not rendering correctly past the normal world limits.
 
 - v1.34
   - Adds a new field `groundOffset` to the `expand_locations.yaml`.